New function template for inviting developers
